+++ HISTORY.md
@@ -1,5 +1,8 @@
 History of features added to LiChess Tools in time (reversed order)
 
+25 Jun
+  - mchat options (the chat used in team and study)
+
 17 Jun
   - extract FENs in PGN Editor